Move round save to event handler, remove useEffect
Per React "You Might Not Need an Effect": saving the round is triggered
by a specific user action (clicking AIKA PÄÄTTYI!), not by an external
system, so the fetch belongs in the click handler rather than watching
completedTime in an effect.

Moves the POST /api/rounds call into handleCompleteRound() in timer.tsx,
computes completedAt once and passes it to both the Liveblocks mutation
and the fetch. Removes all effect/ref code from MysteryRoundPage.

This also guarantees exactly one DB write per round end (the client who
clicked), instead of every connected client racing behind a ref guard.
